mabdullahz/LP-1902: [White labelling] Changes required in partner registration flow (#1426)
* Moved g2a forms and views to partners

* Moved tests and all views outside to partners

* Add required fields in Partner and UserPartner model

* Added fields and made migration required

* Added the USER_LIMIT check during registration

* Add form override function and fix broken tests

* Fix unit test doc strings for updated tests

* Incorporate review changes

* Add missing USER_LIMIT field validations

* Fixed the JSON field single quotes error
