--- openedx/features/partners/give2asia/__init__.py
@@ -1,2 +1,0 @@
-# The name of the directory `give2asia` should must match
-# with the slug of some partner in admin/partners/partner/
